Use UUID instead of PID
When multinple tests are run under py.test, all the docker
images are getting the same name as the getpid() is returning
the same value for them. Changing the code to use UUID instead of
PID. This will ensure unique name across the docker images.

Some of the OVS commands are printing multiple new line
charecters. Adding a string replace command for the OVS output.

Change-Id: Ia7df3fdb0c25b1077910961da7fb7946f597ddf3
Signed-off-by: Sreedhar <sreedharr@hp.com>