Refactor container name logic
In newer Docker versions '_' is now forbidden as a hostname.
Currently the container naming logic is all over the place so we will
have to embark on a journey to bring this to a sane state where we can
change the container naming format.

Furthermore, this change makes the logging more likely to work across
distributions by looking at file existance instead of distro name.
It also makes the log messages a bit more verbose.

Finally, this change aboards the test straight away if the
'docker run' call failed. The old behavior was to wait 300 seconds
and then time out.

Change-Id: I96b1e37416a09527f0094803e3a1b152018056cb
Signed-off-by: Christian Svensson <blue@cmd.nu>